globalized types and cleaned up type for historical metrics
