+++ server/Controllers/getHistoricalPrometheusData.ts
@@ -22,6 +22,9 @@ export const getHistoricalPrometheusData = async (pod: string, query: string) =>
     const info:any = queryResult.json();     //[1234.12, 1234]
     console.log(info);
     console.log(info.data.result[0].values)
+    for (let i = 0; i < info.data.result[0].values.length; i++) {
+      info.data.result[0].values[i][1] = Number(info.data.result[0].values[i][1]);
+    }
     return info.data.result[0].values; 
   }
   catch {
