--- server/types-server.tsx
@@ -1,22 +1,0 @@
-
-// --------------------- Types for Pods in getPods.ts -------------------------------
-// interface PodNode {
-//   pod: string
-//   node: string
-// }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-// module.exports = {PodNode};
